fix(security): an image cannot claim enough pixels to take the API down
The thumbnail job handed presigned uploads straight to image.Decode, which
reserves memory for every pixel the header claims. A 72-byte PNG claiming
8000x8000 allocated 244 MB; at 30000x30000 it asks for 3.4 GB. The job runs in
the API process and was retried five times.

- storage image helpers read at most 50 MB, check width x height against the
  media profile's MaxPixels from the header, then decode; they return
  ErrImageTooLarge or ErrUnreadableImage
- the thumbnail job returns asynq.SkipRetry for both, so the task is archived
  after one attempt
- the media pipeline's pixel checks multiply in int64
- storage/image_test.go ships a 30000x30000 PNG and fails if refusing it
  allocates over 64 MB
- grit upgrade updates jobs/workers.go after delivering storage

+++ internal/scaffold/api_media_files.go
@@ -432,7 +432,7 @@ func Transform(r io.ReadSeeker, p Profile) (Result, error) {
 	if err != nil {
 		return Result{}, fmt.Errorf("reading image header: %w", err)
 	}
-	if px := cfg.Width * cfg.Height; px > p.MaxPixels {
+	if px := int64(cfg.Width) * int64(cfg.Height); px > int64(p.MaxPixels) {
 		return Result{}, fmt.Errorf(
 			"image is %dx%d (%d megapixels), over the %d megapixel limit",
 			cfg.Width, cfg.Height, px/1000000, p.MaxPixels/1000000)
@@ -926,7 +926,7 @@ func Transform(r io.ReadSeeker, p Profile) (Result, error) {
 	}
 	srcW, srcH := probe.Width(), probe.Height()
 	probe.Close()
-	if px := srcW * srcH; px > p.MaxPixels {
+	if px := int64(srcW) * int64(srcH); px > int64(p.MaxPixels) {
 		return Result{}, fmt.Errorf(
 			"image is %dx%d (%d megapixels), over the %d megapixel limit",
 			srcW, srcH, px/1000000, p.MaxPixels/1000000)
